QueueRequestOptions setter
