+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/AzureStorageClient.java
@@ -14,6 +14,7 @@ import com.microsoft.azure.storage.blob.CloudBlobContainer;
 import com.microsoft.azure.storage.blob.CloudBlockBlob;
 import com.microsoft.azure.storage.queue.CloudQueue;
 import com.microsoft.azure.storage.queue.CloudQueueMessage;
+import com.microsoft.azure.storage.queue.QueueRequestOptions;
 import com.microsoft.azure.storage.table.CloudTable;
 import com.microsoft.azure.storage.table.TableOperation;
 import com.microsoft.azure.storage.table.TableServiceEntity;
@@ -47,14 +48,14 @@ class AzureStorageClient {
         this.operationContext = IngestionUtils.httpClientPropertiesToOperationContext(httpClientProperties);
     }
 
-    void postMessageToQueue(String queuePath, String content) throws StorageException, URISyntaxException {
+    void postMessageToQueue(String queuePath, String content, QueueRequestOptions options) throws StorageException, URISyntaxException {
         // Ensure
         Ensure.stringIsNotBlank(queuePath, "queuePath");
         Ensure.stringIsNotBlank(content, "content");
 
         CloudQueue queue = new CloudQueue(new URI(queuePath));
         CloudQueueMessage queueMessage = new CloudQueueMessage(content);
-        queue.addMessage(queueMessage, 0, 0, null, this.operationContext);
+        queue.addMessage(queueMessage, 0, 0, options, this.operationContext);
     }
 
     void azureTableInsertEntity(String tableUri, TableServiceEntity entity) throws StorageException,
