+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/ManagedStreamingIngestClient.java
@@ -45,7 +45,7 @@ public class ManagedStreamingIngestClient implements IngestClient {
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     public static final int ATTEMPT_COUNT = 3;
     public static final int MAX_STREAMING_SIZE_BYTES = 4 * 1024 * 1024;
-    final QueuedIngestClient queuedIngestClient;
+    final QueuedIngestClientImpl queuedIngestClient;
     final StreamingIngestClient streamingIngestClient;
     private final ExponentialRetry exponentialRetryTemplate;
 
@@ -103,7 +103,7 @@ public class ManagedStreamingIngestClient implements IngestClient {
             @Nullable HttpClientProperties properties)
         throws URISyntaxException {
         ConnectionStringBuilder dmConnectionString = new ConnectionStringBuilder(engineConnectionString);
-        dmConnectionString.setClusterUrl(QueuedIngestClient.generateDmUriSuggestion(new URIBuilder(engineConnectionString.getClusterUrl())));
+        dmConnectionString.setClusterUrl(QueuedIngestClientImpl.generateDmUriSuggestion(new URIBuilder(engineConnectionString.getClusterUrl())));
         return new ManagedStreamingIngestClient(dmConnectionString, engineConnectionString, properties);
     }
 
@@ -116,7 +116,7 @@ public class ManagedStreamingIngestClient implements IngestClient {
             ConnectionStringBuilder engineConnectionStringBuilder,
             @Nullable HttpClientProperties properties) throws URISyntaxException {
         log.info("Creating a new ManagedStreamingIngestClient from connection strings");
-        queuedIngestClient = new QueuedIngestClient(dmConnectionStringBuilder, properties);
+        queuedIngestClient = new QueuedIngestClientImpl(dmConnectionStringBuilder, properties);
         streamingIngestClient = new StreamingIngestClient(engineConnectionStringBuilder, properties);
         exponentialRetryTemplate = new ExponentialRetry(ATTEMPT_COUNT);
     }
@@ -125,7 +125,7 @@ public class ManagedStreamingIngestClient implements IngestClient {
             AzureStorageClient storageClient,
             StreamingClient streamingClient) {
         log.info("Creating a new ManagedStreamingIngestClient from raw parts");
-        queuedIngestClient = new QueuedIngestClient(resourceManager, storageClient);
+        queuedIngestClient = new QueuedIngestClientImpl(resourceManager, storageClient);
         streamingIngestClient = new StreamingIngestClient(streamingClient);
         exponentialRetryTemplate = new ExponentialRetry(ATTEMPT_COUNT);
     }
@@ -135,7 +135,7 @@ public class ManagedStreamingIngestClient implements IngestClient {
             StreamingClient streamingClient,
             ExponentialRetry retryTemplate) {
         log.info("Creating a new ManagedStreamingIngestClient from raw parts");
-        queuedIngestClient = new QueuedIngestClient(resourceManager, storageClient);
+        queuedIngestClient = new QueuedIngestClientImpl(resourceManager, storageClient);
         streamingIngestClient = new StreamingIngestClient(streamingClient);
         exponentialRetryTemplate = retryTemplate;
     }
